:wrench: rename types to front-chat-types
