+++ lib/types/index.ts
@@ -1,3 +1,3 @@
-import * as types from './types';
+import * as types from './front-chat-types';
 
 export {types};
